requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
numpy==2.1.3
langchain==0.3.7
langchain-openai==0.2.6
openai==1.54.3
wikipedia==1.4.0
pandas==2.2.3
matplotlib==3.9.2
tqdm==4.66.5
pytest==8.3.3
rich==13.9.4
